requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
dnspython==2.6.1
docstring_parser==0.16
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-text-splitters==0.2.1
langsmith==0.1.77
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.5
packaging==24.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pymongo==4.7.3
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
shapely==2.0.4
six==1.16.0
SQLAlchemy==2.0.30
tenacity==8.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.1
vertexai==1.49.0
Werkzeug==3.0.3
yarl==1.9.4
psycopg2==2.9.6